requirements.txt
robotframework
robotframework-browser
robotframework-faker
pymongo
bcrypt